Trait snowbridge_ethereum_beacon_client::pallet::Config

pub trait Config: Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type ForkVersions: Get<ForkVersions>;
    type MaxExecutionHeadersToKeep: Get<u32>;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

type ForkVersions: Get<ForkVersions>

type MaxExecutionHeadersToKeep: Get<u32>

Maximum number of execution headers to keep

type WeightInfo: WeightInfo